Task1 updated from feedback
